breaking stuff out in different components, added higscore handling
