+++ src/nameform-component.js
@@ -1,0 +1,34 @@
+import React from 'react'
+
+export class NameFormComponent extends React.Component {
+  constructor(props) {
+    super(props)
+    this.state = {value: ''}
+
+    this.handleChange = this.handleChange.bind(this)
+    this.handleSubmit = this.handleSubmit.bind(this)
+  }
+
+  handleChange(event) {
+    this.setState({value: event.target.value})
+  }
+
+  handleSubmit(event) {
+    event.preventDefault();
+    if (this.props.onNameSubmitted) {
+      this.props.onNameSubmitted(this.state.value)
+    }
+  }
+
+  render() {
+    return (
+      <form onSubmit={this.handleSubmit}>
+        <label>
+          Name:
+          <input type="text" value={this.state.value} onChange={this.handleChange} />
+        </label>
+        <input type="submit" value="Submit" />
+      </form>
+    );
+  }
+}
